Moved test_gpr.py and test_kernels.py to tests folder
